+++ packages/deprecated/facts/package.js
@@ -1,11 +1,11 @@
 Package.describe({
   summary: "Publish internal app statistics",
-  version: '1.0.0'
+  version: '1.0.9'
 });
 
 Package.onUse(function (api) {
-  api.use('ecmascript', ['client', 'server']);
-  api.use('underscore', 'server');
+  api.use(['underscore'], ['client', 'server']);
+  api.use(['templating@1.2.13', 'mongo', 'ddp'], ['client']);
 
   // Detect whether autopublish is used.
   api.use('autopublish', 'server', {weak: true});
@@ -14,8 +14,8 @@ Package.onUse(function (api) {
   // us.
   api.use('ddp', 'server', {unordered: true});
 
-  api.mainModule('facts_base_server.js', 'server');
-  api.mainModule('facts_base_common.js', 'client');
+  api.addFiles('facts.html', ['client']);
+  api.addFiles('facts.js', ['client', 'server']);
 
   api.export('Facts');
 });