+++ packages/facts-ui/facts_ui_client.js
@@ -1,15 +1,9 @@
 import { Facts, FACTS_COLLECTION, FACTS_PUBLICATION } from 'meteor/facts-base';
 
-console.log('Facts UI client');
-
 Facts.server = new Mongo.Collection(FACTS_COLLECTION);
 
 Template.serverFacts.helpers({
-  factsByPackage: function () {
-    const facts = Facts.server.find();
-    console.log("Facts", facts);
-    return facts;
-  },
+  factsByPackage: () => Facts.server.find(),
   facts: function () {
     const factArray = [];
     _.each(this, function (value, name) {